Move config to base dir
