+++ tests/conftest.py
@@ -25,7 +25,6 @@ from omop_emb.backends.faiss import FaissEmbeddingBackend
 from omop_emb.backends.pgvector import PGVectorEmbeddingBackend
 from omop_emb.backends.registry import ensure_model_registry_schema, ModelRegistry
 from omop_emb.interface import EmbeddingInterface
-from omop_emb.backends.config import IndexType
 
 
 TEST_DB_NAME = os.getenv("TEST_DATABASE_NAME", "test_omop_emb")
