+++ tests/shared_backend_tests.py
@@ -3,7 +3,7 @@
 import pytest
 import numpy as np
 
-from omop_emb.backends.config import IndexType, MetricType
+from omop_emb.config import IndexType, MetricType
 from omop_emb.backends.embedding_utils import EmbeddingConceptFilter
 from omop_emb.backends.errors import ModelRegistrationConflictError
 from .conftest import CONCEPTS, MODEL_NAME, EMBEDDING_DIM, TEST_CONCEPT_EMB
